month 1.0

+++ lib/page/bookKeeping/keeping_income.dart
@@ -189,7 +189,7 @@ class _KeepIncomeState extends State<KeepIncome> {
                 double _keepText =
                 double.parse(_keepTextController.text) == 0
                     ? 0
-                    : int.parse(_keepTextController.text);
+                    : double.parse(_keepTextController.text);
                 String _beiZhuText = _beiZhuTextController.text == ''
                     ? '无'
                     : _beiZhuTextController.text;
